refactor: categories component
